Put user controllers into one controller
