+++ src/main/java/edu/kpi/testcourse/rest/UsersController.java
@@ -19,17 +19,18 @@ import io.micronaut.security.authentication.UsernamePasswordCredentials;
 import io.micronaut.security.rules.SecurityRule;
 import io.micronaut.security.token.jwt.render.BearerAccessRefreshToken;
 import java.util.Map;
+import java.util.UUID;
 import javax.inject.Inject;
 
 /**
  * Controller for login.
  */
-@Secured(SecurityRule.IS_ANONYMOUS)
 @Controller("/users")
-public class UserSignInController {
+public class UsersController {
   @Inject
   @Client("/")
   RxHttpClient client;
+
   /**
    * in field "body" in Postman you should provide user email and password like
    * {
@@ -37,7 +38,7 @@ public class UserSignInController {
    *   "password" : "yourPassword"
    * }.
    */
-
+  @Secured(SecurityRule.IS_ANONYMOUS)
   @Post(value = "/signin",
       consumes = MediaType.APPLICATION_JSON,
       produces = MediaType.APPLICATION_JSON)
@@ -69,6 +70,35 @@ public class UserSignInController {
     }
 
     return httpResponse;
+  }
+
+  @Secured(SecurityRule.IS_AUTHENTICATED)
+  @Post(value = "/signout")
+  public HttpResponse signOut() {
+    return HttpResponse.noContent();
+  }
+
+  /**
+   * in field "body" in Postman you should provide user email and password like
+   * {
+   *   "email" : "yourEmail",
+   *   "password" : "yourPassword"
+   * }
+   * then you will see response that your user is created.
+   */
+  @Secured(SecurityRule.IS_ANONYMOUS)
+  @Post(value = "/signup",
+    consumes = MediaType.APPLICATION_JSON,
+    produces = MediaType.APPLICATION_JSON)
+  public HttpResponse signUp(@Body JSONObject object) {
+    User user = Main.getGson().fromJson(object.toJSONString(), User.class);
+    user.setUuid(UUID.randomUUID().toString());
+    boolean response = UserActions.createUser(user);
+    if (response) {
+      return HttpResponse.created("User with email \"" + user.getEmail() + "\" was created.");
+    } else {
+      return HttpResponse.badRequest("User with email \"" + user.getEmail() + "\" already exists.");
+    }
   }
 }
